+++ lib/wifi_connector_plus.dart
@@ -10,6 +10,7 @@ export 'src/models/wifi_connect_result.dart';
 export 'src/models/wifi_credentials.dart';
 export 'src/wifi_qr_parser.dart';
 export 'src/widgets/wifi_qr_scanner_view.dart';
+export 'src/widgets/wifi_qr_scanner_controller.dart';
 
 class WifiConnectorPlus {
   /// Get the current platform version.